requirements.txt
annotated-types==0.7.0
anyio==4.7.0
click==8.1.7
databases==0.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
pydantic==2.10.3
pydantic[email]
pydantic_core==2.27.1
python-jose
passlib[bcrypt]
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
